requirements.txt
Django >4.2, <4.3
asgiref
sqlparse
tzdata
pillow
django-robots
django-taggit